Sca top_n changed and  ugly for loops are here to stay
sca produces variable number of interactionsdepending on dilution. Changed top_n to be lower so that at every dilution the same number of top ranks are retrieved.

Ugly for loops merge the formatting to fold dilution data in with original data. They are uglier than mapply but also work if the format of the dilution data has gaps in it. Mapply just messes with the whole data set in that case. So the ugly for loops stay for now