Liana_wrap does not produce the same results as call_x functions for sca and cellchat
I don't know why but as the new code snippet shows, liana_wrap and call_x functions for sca and cellchat produce different results even when referring to identical resources, identical methods and identical data. The call functions are consistently smaller.

On the upside, using the call_X functions, filtered OP actually DOES produce the same output as the normal OP.

Cleaned up the code a bit to reflect this.