Modified call_x function calls, added section to filter OP before dilution
To ensure the same standards are being used in the liana_wrap function as the call_x functions I set the defaults for interaction filtering in the call_x functions to be the same as in liana_wrap. Practically, liana_wrap and call_x will now produce the same output in the code, as opposed to before where the call_x functions truncated the results according to their own default parameters. This resolves strange behaviour I was seeing in the Snippet code from yesterday.

Based on the changes in result nrows top n could be updated upwards a bit. Not really relevant.

Also, now that I've confirmed that a filtered OP resource produces the same outputs as a non-filtered one, I'vee added a filtration step for OP before dilutions are made. For more info check the comment in the code.

Replaced an mapply with an ugly for loop, because the loop is more reliable.