[MAJOR] Compatible with newest LIANA version, natmi added
In the newest version of LIANA you must refer to methods with the name 'call_*' when you want to use the original. Since we are benchmarking using a reimplementation is not as good as the original so all mentions were changed to not include reimplementations. The newest version of LIANA doesn't have a cellchat reimplementation, so that stays the same.

Replaced individual calls to  methods with a 'methods_vector', this is a step in the direction of making the script method-toggleable, but is by far not there yet.

Implemented natmi across the board. Natmi has some problematic behaviour when used with liana wrap where previous results contaminate current ones but that is already fixed in this version by making the method save to a custom folder each time.

Now that liana_wrap can handle custom resources all individual call_* functions have been replaced with liana_wrap calls, which greatly compacted section E. All the corect parameters are passed to liana_wrap and the outputs are all the same as the old call_* functions (extensively tested).

Minorly tweaked some formatting. More formatting fixes upcoming.