[MAJOR] Warnings revamped and function output that describes results
Warnings always in all caps.

Added a portion that analyses the outputs in some qc measures of network topology to make sure the function ran as intended. If verbose == TRUE the full detailed output will be printed for the user and can help verify everything went well in the dilution.

Some automatic basic checks run in the background and throw warnings even if verbose == FALSE if the topology or other parameters are very wrong.