[MAJOR] Description, Readability, Subfunctions, Args rework, Output rework, Small Bugs
Added a detailed description to dilute_Resource() that describes the general purpose of the function and of dilution, not just the parameters.

Added a bunch of comments and curly-bracket-structuring to make the code far easier to understand and more readable. Did a bunch of reordering, reformatting, etc.

Made dilute_Resource() into more of a handler than a standalone function. dilute_Resource protects the top rank interactions and finds the subset of rows of resource to dilute, as well as analysing the diluted resource to make sure its safe. The actual dilution methods and extraction of gene names from the data set is now passed off on to new functions. extract_unconflicitng_genes() gets the genes, random_Dilute and preserve_Dilute are the two currently available dilution methods. Each of  these own functions has its own documentation and description, and are well commented throughout as well.

Reworked some function arguments of dilute_Resource(9. Namely, top_rank_df is now reduced to a vector or list of LR_Pairs, since that's all that was ever needed from top_rank_df. In addition, dilution_feature_type was renamed to feature_type since its less wordy and should be clear in the context of the main dilution function.

Minor things:
- Output had some slight tweaks, it now explains what semi-preserved topology means
- Warnings now come after the output
- Fixed issue where warning_thrown and error_thrown would incorporate r_topology_okay when random_Dilute() hadn't even been used, or p_topology_okay when preserve_Dilute() hadn't been used.