[MINOR] Compatibility with new functions, updated plotting, updated script_params
Changed arguments of dilute_Resource() to work with its updated version.

The plots now get a  note on the dilution type in the plot subtitle.

Script params now saves sink_output and the save path to the log file in itself, as well as preserve_topology. These values are also removed from the general environment to declutter it.