Functions take seeds
In order to calculate multiple permutations of dilution and aggregate results, all the dilution function must be able to take a seed for their elements of randomness.

Each of them now takes the arg master_seed. Every time a seed is needed for some random operation, the seed is set to a unique seed derivative of master_seed. As long as the different seeds passed to master seed are whole numbers, there should be no chance of issue. To  enforce this, reminders were added in the documentation and master seed is floored at the start of every function that uses it.