Removed saving environment functionality, changed script_params arguemnt, removed useless arguments
The data generated by one iteration of this function is not nearly as interesting as the data iterated by many. Saving the environment iteration by iteration is not really a desired feature, so its been removed. This made run_mode and save_results useless as arguments for this function, so they've been removed.

Additionally, returning arguments in script_params that the user already knows (because they passed them to this function) is very strange. Instead, the function now only returns valuable metadate the user doesn't already know. Returning args the user provided is just not necessary here.