Refined sink_output and added a less radical option
Sinking output was made to tackle the fact that the LIANA subfunctions output so much information that relevant warnings and errors fall by the wayside.

Since sinking the output is a bit extreme, a new argument is introduced here, which diverts warnings from LIANA so that relevant non-liana method warnings get to be seen. (LIANA method warnings are often irrelevant and repetitive). This doesn't 100% fix the fact that liana methods produce a ton of output, but it makes the tracking of warnings and errors much easier.