Logs savepath is now user specified
The user now specifies under which file path sunk logs and warning logs are saved, and the automatic construction of a save path has been moved to the handler. This way all the iterations of one lapply call get to save their logs at one unique location that no further lapply call will modify (since the time of the lapply call is used as a marker). In the old system, logs run with identical parameters would be appended to old logs, which is confusing.

Since the log save location is an argument for the function, it is no longer returned in metadata, which now serves only to communicate runtime.

The restructuring of metadata has thus changed a bit, and metadata was a specific enough object that its now a sub-object of the script params.

Metadata is in this case data that is about the dilution_Robustness() call that the user didn't know when running lapply or something that would lead them to something they didn't know when running the lapply.