Function log file paths are now user defined arguments
The function no longer creates automatic filepaths itself, they are now user supplied. The automatic creation couldn't group logs for multiple iterations sensibly, and when run with the same paramters on the same day new logs culd just be appended to unrelated older logs.

Now the user supplies the filepath, they'll know where to find the log and it is their responsibility to name the log in such a way that the name isn't redundant between runs.

The handler for example can know when the lapply iteration of this function is run, which can be used to create a unique log name that stores all the iterations within it.

Since filepaths are now user provided, they are no longer returned as metadata.