Updates resources_OP structure
Every method uses the same top_ranks to dilute and thus generates the same resource. Since every resource (per dilution step) is identical, there's no need to store one per method, since the methods no longer make a difference. The code now reflects this, and is simpler because of it.