No more script params with dilution_Robustness
Script_params is a problematic object because it stores important information as an intermediate object in the environment.

In future, only the analysis results should be stored in the environment.

The user can define the parameters to run the script with by changing the default settings of the functions in the iterator_functions script.

The overall project is a work in progress, this represents the completion of removing script_params and making resource_robustness() still work.