Shifted runtime calculation and metadata curation
Runtime calculations now take place through a new function, and are shifted as to where they occur to make more sense.

In addition, a new function called summarise_Metadata now produces a metadata object instead of doing it manually. At the same time, summarise_Metadata stores important arguments for the iterator within its default arguments, allowing it to be a reference point for other functions that need to know if this is a trial run for example, and whether or not to save the results.