The generation of these parameters has changed
Time_of_run must be generated flexibly, and therefore can't be a function default.

Master_seed list can't conveniently be a parameter, having to define every element of the list as a default argument in a function would be very tedious.

Thus, even though these parameters are relevant to the running of the iterator, they can't solely be defined as function defaults, they need to be their own objects free flaoting in the environment.

However, it is possible to shift how these parameters are generated. By having them generated by a function, the entire definition of parameters is now in the iterator parameter script, and the basis for master_seed_list can be set like any other parameter in the defaults in iterator params.