[MAJOR] USING FUNCTION DEFAULTS IS BAD
Instead of storing iterator parameters inside function defaults, the entire iterator should be a function on its own that gets passed certain arguments by the user and then runs on those defaults.

The entire project of removing script params and changing to function defaults with  formals was misguided and based on a misunderstanding. Functionalisation and so on was still useful of course but the next commits will be dealing with undoing iterator paramters, turning the iterator into a function, and adding the correct arguments to the parameter dependent function calls.

As a apart of this, a script is needed where we run the iterator wrapper to receive results, which is what this addition represents.