Converted into Function, updated parameters. Especially output parameters.
Converted the iterator into a big wrapper function of the old iterator.

Each parameter has a default, so the user only needs to pass the arguments that deviate from the defaults to run the iterator.

Since the iterator has certain outputs, and resource_Robustness has certain outputs, I added a parameter for each of the outputs, bundled for the resource_Robustness() output and master for the iterator output.  Changed code to reflect this. Added segment that compiles iterator_results and returns them.