Updated paramter creation for parameters that couldn't be saved as defaults
There used to be parameters that couldn't be directly saved as function defaults, they had to be editable objects in the environment. Turns out, that's not a downside, you just want all the arguments to be passed by a wrapper from a higher scope, which we've implemented now.

So just like all the other arguments these will also be free floating. There's no reason to generate these with a function, so they just have a section of their own where they are generated for the iterator and run. Deletec the old function.