Modified printing outputs and changed reshuffler iteration
The reshuffler used to map by seed and then by mismatch proportion, this order has been reversed (map by mismatch, within that, map by seed). This was done because calculating multiple different samples (i.e. different seeds) at the same mismatch proportion seemed more intuitive, and because the mismatch proportion was a more informative high-level structure for the list of results to have. It is easier to get the values you need when the resuts are ordered by 0, 40, 80 % then by Seed 1, 2, 3.

The printed output changes just made use of the already existing print_Title function to save unnecessary lines of code and repetition.