Added iteration of LIANA++ over all cluster annotations
Added a wrapper function that runs LIANA++ with the appropriate warnings.

Added a function that calculates the LIANA++ results for the default cluster annotations and calculates LIANA for every reshuffled cluster annotation (iterating over mismatch proportions and seeds). Also works with runtime in a limited way and appends the default results and reshuffled results together.

Back in the iterator, the runtime is then saved separately and the results are transposed a few times to facilitate with future processing.

Also, where the warnings was saved is output to the user.