Added comparison of top ranked interactions
We add a methods list to map over, within every map we evaluate if the method is cellchat. Depending on the method we then implement a function to first get top ranks, next format them to have an LR_Pair and LR_ID and then compare the overlap between the original top ranks and later predicted ones.

Added the format_top_ranks method that adds LR_Pairs and LR_IDs to top ranks tibbles.