Added code for when errors occur in LIANA
If LIANA return an error in the iteration process, this basically means a critical fail for the method. If it is not a trial run, the results (+ clusters + metadata) are bundled up and saved to the hardrive, marked with ERROR (for forensic bug checking) and the function will be unable to calculate any further without liana_wrap.

If it is a trial run, errors from liana will be filtered out, replaced with NA's and the code moves onward with the idea that the error was probably caused by a method being unable to handle higher degrees of cluster dilution.

If the use wants to ignore LIANA errors that occur in a serious run, they'll have to enable that by hand.