Fixed issues from inconsistent metadata formatting
Metadata and especially cluster annotations aren't consistent between testdatas.  In this update, we added a column to the metadata of testdata called cluster_key that labels the clusters from 1 to n, and then sets the idents of testdata to be equal to cluster_key. This makes the cluster labelling across the board identical, regardless of the testdata. This alos makes the old cluster_col parameter obsolete, which is removed everywhere where it is mentioned.